+++ inc/Options/Fields/CheckboxField.php
@@ -39,7 +39,7 @@ class CheckboxField extends InputField {
 	}
 
 	protected function default_validate( mixed $value ): bool {
-		$required = isset( $this->args['required'] ) && $this->args['required'];
+		$required = $this->is_required();
 
 		if ( $required && ! $value ) {
 			$this->add_error( "{$this->title} is required." );