+++ inc/Options/Fields/Field.php
@@ -83,6 +83,18 @@ abstract class Field {
 		}
 	}
 
+	public function is_required(): bool {
+		if ( ! isset( $this->args['required'] ) ) {
+			// If required flag is not set, then by default field is not required
+			return false;
+		}
+		if ( \is_callable( $this->args['required'] ) ) {
+			// phpcs:ignore NeutronStandard.Functions.DisallowCallUserFunc.CallUserFunc
+			return (bool) \call_user_func( $this->args['required'], $this );
+		}
+		return (bool) $this->args['required'];
+	}
+
 	public function is_visible(): bool {
 		if ( ! isset( $this->args['visible'] ) ) {
 			return true;
@@ -203,12 +215,14 @@ abstract class Field {
 	}
 
 	protected function add_error( string $message, string $error_id = '' ): void {
-		add_settings_error(
-			"{$this->group->name()}_errors",
-			$error_id ?: "{$this->id}_error",
-			$message,
-			'error'
-		);
+		if ( $this->is_visible() || $this->is_required() ) {
+			add_settings_error(
+				"{$this->group->name()}_errors",
+				$error_id ?: "{$this->id}_error",
+				$message,
+				'error'
+			);
+		}
 	}
 
 	protected function call_args_callback( string $key, mixed $value, callable $default_callback ): mixed {